separate out the core classes into their independent files
